+++ src/queue/utils/constants.ts
@@ -71,7 +71,8 @@ export const HeaderConstants = {
   ACCESS_CONTROL_ALLOW_HEADERS: "Access-Control-Allow-Headers",
   ACCESS_CONTROL_MAX_AGE: "Access-Control-Max-Age",
   ACCESS_CONTROL_REQUEST_METHOD: "access-control-request-method",
-  ACCESS_CONTROL_REQUEST_HEADERS: "access-control-request-headers"
+  ACCESS_CONTROL_REQUEST_HEADERS: "access-control-request-headers",
+  X_MS_VERSION: "x-ms-version"
 };
 
 export const SECONDARY_SUFFIX = "-secondary";
@@ -87,4 +88,25 @@ export const DEFAULT_QUEUE_PERSISTENCE_ARRAY: StoreDestinationArray = [
     persistencyPath: DEFAULT_QUEUE_PERSISTENCE_PATH,
     maxConcurrency: 1
   }
+];
+
+export const ValidAPIVersions = [
+  "2019-02-02",
+  "2018-11-09",
+  "2018-03-28",
+  "2017-11-09",
+  "2017-07-29",
+  "2017-04-17",
+  "2016-05-31",
+  "2015-12-11",
+  "2015-07-08",
+  "2015-04-05",
+  "2015-02-21",
+  "2014-02-14",
+  "2013-08-15",
+  "2012-02-12",
+  "2011-08-18",
+  "2009-09-19",
+  "2009-07-17",
+  "2009-04-14"
 ];